requirements.txt
ipykernel
python-dotenv
fastapi
uvicorn
pydantic
pydantic[email]
pytest
pytest-asyncio
httpx
python-dotenv
pytest-dotenv
sqlalchemy
psycopg2-binary
pydantic-settings
fastapi-pagination
passlib[bcrypt]
python-jose[cryptography]